# docs/swagger.yaml
basePath: /api
definitions:
...
  helpers.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      deleted_by:
        type: integer
      email:
        type: string
      emergency_contact:
        type: string
      id:
        type: integer
      identity_number:
        type: string
      identity_photo_url:
        type: string
      identity_type:
        type: string
      is_blacklisted:
        type: boolean
      modified_at:
        type: string
      modified_by:
        type: integer
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact:
    email: developer.haritsnb@gmail.com
    name: Harits Nala Barrun
  description: Backend RESTful API Sistem Kasir & Manajemen Rental Alat Camping menggunakan
    Golang + Gin + PostgreSQL + JWT Blacklist + KYC Upload.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Outdoor Gear Rental POS API
  version: "1.0"
paths:
...
  /customers:
    get:
      description: Memuat seluruh daftar pelanggan terdaftar beserta status KYC dan
        blacklist
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Customer'
                  type: array
              type: object
        "401":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Daftar Semua Pelanggan
      tags:
      - Customers
    post:
      consumes:
      - multipart/form-data
      description: Mendaftarkan data pelanggan baru dan mengunggah foto kartu identitas
        (KTP/SIM/Paspor)
      parameters:
      - description: Nama Lengkap
        in: formData
        name: name
        required: true
        type: string
      - description: Tipe Identitas (KTP/SIM/Paspor)
        in: formData
        name: identity_type
        required: true
        type: string
      - description: Nomor Unik Identitas
        in: formData
        name: identity_number
        required: true
        type: string
      - description: Nomor HP / WA Aktif
        in: formData
        name: phone
        required: true
        type: string
      - description: Alamat Domisili
        in: formData
        name: address
        required: true
        type: string
      - description: Kontak Darurat
        in: formData
        name: emergency_contact
        type: string
      - description: Email Pelanggan
        in: formData
        name: email
        type: string
      - description: Catatan Tambahan
        in: formData
        name: notes
        type: string
      - description: Foto Fisik Kartu Identitas
        in: formData
        name: identity_photo
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Customer'
              type: object
        "400":
          description: Bad Request / Upload Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request / Upload Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Bad Request / Upload Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Registrasi Pelanggan & Upload KYC
      tags:
      - Customers
  /customers/{id}:
    delete:
      description: Menonaktifkan data pelanggan (Admin Only)
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Soft Delete Pelanggan
      tags:
      - Customers
    get:
      description: Memuat data lengkap pelanggan dan URL foto identitasnya
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Customer'
              type: object
        "401":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Detail Pelanggan Berdasarkan ID
      tags:
      - Customers
    post:
      description: Memulihkan kembali data pelanggan yang berstatus soft-deleted (Admin
        Only)
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Restore Pelanggan
      tags:
      - Customers
    put:
      consumes:
      - multipart/form-data
      - application/json
      description: Memperbarui biodata pelanggan dan opsional mengganti file foto
        kartu identitas
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Nama Lengkap
        in: formData
        name: name
        type: string
      - description: Tipe Identitas
        in: formData
        name: identity_type
        type: string
      - description: Nomor Identitas
        in: formData
        name: identity_number
        type: string
      - description: Nomor HP
        in: formData
        name: phone
        type: string
      - description: Alamat
        in: formData
        name: address
        type: string
      - description: Kontak Darurat
        in: formData
        name: emergency_contact
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Catatan
        in: formData
        name: notes
        type: string
      - description: Foto Identitas Baru
        in: formData
        name: identity_photo
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Pelanggan (Flexible Partial + Opsional Foto)
      tags:
      - Customers
  /customers/{id}/blacklist:
    patch:
      consumes:
      - application/json
      description: Memasukkan atau menghapus pelanggan dari daftar blacklist (Admin
        Only)
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload Blacklist
        in: body
        name: request
        required: true
        schema:
          properties:
            is_blacklisted:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Ubah Status Blacklist Pelanggan
      tags:
      - Customers
  /customers/{id}/force:
    delete:
      description: Menghapus fisik data pelanggan dari database (Admin Only)
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: Bad Request / Foreign Key Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: Bad Request / Foreign Key Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "403":
          description: Bad Request / Foreign Key Error
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Permanent Pelanggan (Force Delete)
      tags:
      - Customers
  /customers/deleted:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Sukses
          schema:
            allOf:
            - $ref: '#/definitions/helpers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Customer'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Pelanggan Terhapus (Trash)
      tags:
      - Customers
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"